backend/requirements.txt
fastapi
uvicorn[standard]
numpy
rasterio
Pillow
matplotlib